+++ src/app/(router)/(main)/favicon.ico
